Modifications tableau des scores

+++ Implementation/Implementation_AG_Propre/Sources/initialisation_population.cpp
@@ -11,7 +11,7 @@ int Individu::tailleIndividu = 0;
 int Individu::nombreCritere = 0;
 float Individu::probaMutation = 0;
 int* chromosome = NULL;
-int* score = NULL;
+float* score = NULL;
 int* rang = NULL;
 
 // Constructeurs //
@@ -45,10 +45,10 @@ Individu::Individu(float donnees[3]){	// ça ok Ce constructeur crée un chromos
 		chromosome[i] = rand()%2;
 	}
 	//// Initialisation de score et rang ////
-	this->score =(int*) malloc(this->nombreCritere*sizeof(int));
+	this->score =(float*) malloc(this->nombreCritere*sizeof(float));
 	this->rang =(int*) malloc(this->nombreCritere*sizeof(int));
 	for(int i = 0; i < this->nombreCritere; i++){
-		this->score[i] = 0;
+		this->score[i] = 0.0;
 		this->rang[i] = 0;
 	}
 }
@@ -58,7 +58,7 @@ Individu::Individu(Individu &indiv){
 	for(int i=0;i<tailleIndividu+1;i++){
 		this->chromosome[i]=indiv.getChromosome()[i];
 	}
-	this->score =(int*) malloc(this->nombreCritere*sizeof(int));
+	this->score =(float*) malloc(this->nombreCritere*sizeof(float));
 	this->rang =(int*) malloc(this->nombreCritere*sizeof(int));
 }
 
@@ -79,7 +79,7 @@ Individu::~Individu() {
 int* Individu::getChromosome(){ 		// ça ok
 	return this->chromosome; 
 }
-int Individu::getScore(int i){ 			// ça ok
+float Individu::getScore(int i){ 			// ça ok
 	return this->score[i];
 }
 int Individu::getRang(int i){ 			// ça ok
@@ -103,7 +103,7 @@ void Individu::setChromosome(int* chromosome2){
 	this->chromosome = chromosome2;
 	// Non pas comme ça //
 }
-void Individu::setScore(int score, int i){
+void Individu::setScore(float score, int i){
 	this->score[i] = score;
 }
 void Individu::setRang(int rang, int i){
